pyproject.toml
[project]
name = "p2c"
version = "0.1.0"
description = ""
authors = [
    {name = "Clément Harang",email = "clement.harang22@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-statemachine (>=2.5.0,<3.0.0)",
    "prettytable (>=3.16.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
